requirements.txt
django
requests
pytest-django
